requirements.txt
Flask==3.0.2
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-generativeai==0.3.0